fix newlib nano _sbrk() is not linked for checking heap
